pyproject.toml
[project]
name = "researcher-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "chandrakant-scrm",email = "chandrakant.shinde@simplecrm.com"}
]
readme = "README.md"
requires-python = "==3.11.11"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "crewai (>=0.102.0,<0.103.0)",
    "crewai-tools (>=0.36.0,<0.37.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "streamlit (>=1.43.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "researcher_demo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
